GKE R361.
Changed use of BaroDiff in UAVPSet to ROCProp for maintaining rate of climb. This is used to maintain a sensible descent rate when dropping from a high altitude to a lower altitude when doing RTH. 

Added a define to double the amount of up compensation each hover adjustment. 

Removed hover trace code causing motor pulsing.

Not flight tested.